+++ flask_hmac/flask_hmac.py
@@ -42,6 +42,15 @@ class Hmac(object):
         self.hmac_disarm = app.config.get('HMAC_DISARM', False)
 
     def auth(self, route):
+        ''' Route decorator
+
+        .. sourcecode:: python
+
+            @app.route("/hmac_auth_view")
+            @hmac.auth  # decorate view
+            def hmac_auth_view():
+                return "hmac_auth_view"
+        '''
         @wraps(route)
         def decorated_view_function(*args, **kwargs):
             try:
@@ -54,15 +63,30 @@ class Hmac(object):
     def abort(self):
         abort(403)
 
-    def _hmac_factory(self, data, digestmod=None):
-        return hmac.new(self.hmac_key, data, digestmod=self.digestmod)
+    def _hmac_factory(self, data, key=None, digestmod=None):
+        key = six.b(key) if key else self.hmac_key
+        return hmac.new(key, data, digestmod=self.digestmod)
 
-    def make_hmac(self, data=''):
-        hmac_token_server = self._hmac_factory(encode_string(data)).digest()
+    def make_hmac(self, data='', key=None):
+        ''' Generates HMAC key
+
+        Arguments:
+            data (str): HMAC message
+            key (str): secret key of another app
+        '''
+        hmac_token_server = self._hmac_factory(encode_string(data), key).digest()
         hmac_token_server = base64.urlsafe_b64encode(hmac_token_server)
         return hmac_token_server
 
     def validate_signature(self, request):
+        ''' Generates HMAC key
+
+        Arguments:
+            request (Request): flask request
+
+        Raise:
+            InvalidSignature: when signatures don't match
+        '''
         if self.hmac_disarm:
             return True
         hmac_token_client = self.get_signature(request)