+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Change Log
 
 0.1.1
 ~~~~~~~~~
+- Able to change secret key in `make_hmac` method
 - Method `validate_signature` created which can be used outside of `Hmac` class
 - Custom exceptions